Add build script and improve/fix tests
