+++ src/zircon/client/no-auto-test
@@ -1,0 +1,1 @@
+Client is not yet implemented, so this cannot be tested.
